requirements.txt
asgiref==3.3.1
Django==3.0.13
django-classy-tags==2.0.0
django-cms==3.7.4
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.4
djangocms-admin-style==1.5.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
easy-thumbnails==2.7.1
fontawesome-free==5.15.3
html5lib==1.1
Pillow==8.1.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.2
webencodings==0.5.1
